chore(models): move SwitchioConfig to enrollment_switchio app
